+++ src/component/habits/list_parts/HabitNameCard.tsx
@@ -1,5 +1,5 @@
 import { Center } from '@chakra-ui/react';
-import { Habit } from '../../interface/main';
+import { Habit } from '../../../interface/main';
 
 const HabitNameCard = ({ name }: Pick<Habit, 'name'>) => {
   return <Center>{name}</Center>;
